refactor: PathFinder 클래스 Component -> Static 변경
