+++ src/main/java/nextstep/subway/path/application/PathService.java
@@ -11,18 +11,16 @@ import org.springframework.stereotype.Service;
 public class PathService {
     private final StationService stationService;
     private final LineService lineService;
-    private final PathFinder pathFinder;
 
-    public PathService(StationService stationService, LineService lineService, PathFinder pathFinder) {
+    public PathService(StationService stationService, LineService lineService) {
         this.stationService = stationService;
         this.lineService = lineService;
-        this.pathFinder = pathFinder;
     }
 
     public ShortestPathResponse findShortestPath(Long startingStationId, Long destinationStationId) {
         Station starting = stationService.findStationById(startingStationId);
         Station destination = stationService.findStationById(destinationStationId);
 
-        return pathFinder.findShortestPath(lineService.findAllLines(), starting, destination);
+        return PathFinder.findShortestPath(lineService.findAllLines(), starting, destination);
     }
 }
